+++ app/views/companion/master/ModeratedSellerExecuteOrder.scala
@@ -12,6 +12,6 @@ object ModeratedSellerExecuteOrder {
     )(Data.apply)(Data.unapply)
   )
 
-  case class Data(orderID: String, gas: Int, password: String)
+  case class Data(orderID: String, gas: Long, password: String)
 
 }
